strip out all spam into seperate package remove references of handler
